feat: refactor repository logic

+++ lib/domain/models/bms_data.dart
@@ -4,7 +4,7 @@ final class BMSData implements IBoatData{
   final List<int> cellsVoltagesMillivolts;
   final List<int> temperatures;
   final double batteryCurrent;
-  final int stateOfCharge;
+  final double stateOfCharge;
   final double totalVoltage;
 
   BMSData({required List<int> voltagesMillivolts, required List<int> temperatures, required this.batteryCurrent, required this.stateOfCharge}) : cellsVoltagesMillivolts = List.unmodifiable(voltagesMillivolts), temperatures = List.unmodifiable(temperatures), totalVoltage = (voltagesMillivolts.fold(0, (previousValue, value) => previousValue + value) / 1000.0);
